requirements.txt
pyserial~=3.5
pycrc
roboticstoolbox-python
scipy==1.11.0
opencv-python~=4.9.0.80
streamlit~=1.34.0
pandas~=2.2.2
numpy~=1.26.4
PyQt5~=5.15.10
roboclaw~=1.0.0.dev1